Re-structures ML job link, cleans up discover link changes
