+++ x-pack/plugins/apm/public/components/shared/Links/DiscoverLinks/__test__/DiscoverLinks.integration.test.tsx
@@ -4,15 +4,13 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { History, Location } from 'history';
+import { Location } from 'history';
 import React from 'react';
-import { render, waitForElement } from 'react-testing-library';
 import { APMError } from '../../../../../../typings/es_schemas/ui/APMError';
 import { Span } from '../../../../../../typings/es_schemas/ui/Span';
 import { Transaction } from '../../../../../../typings/es_schemas/ui/Transaction';
-import { LocationProvider } from '../../../../../context/LocationContext';
 import * as savedObjects from '../../../../../services/rest/savedObjects';
-import { tick } from '../../../../../utils/testHelpers';
+import { getRenderedHref } from '../../../../../utils/testHelpers';
 import { DiscoverErrorLink } from '../DiscoverErrorLink';
 import { DiscoverSpanLink } from '../DiscoverSpanLink';
 import { DiscoverTransactionLink } from '../DiscoverTransactionLink';
@@ -22,30 +20,6 @@ jest
   .mockReturnValue(
     Promise.resolve({ id: 'apm-index-pattern-id' } as savedObjects.ISavedObject)
   );
-
-async function renderLink(
-  Component: React.FC,
-  { search = '' }: { search: string }
-) {
-  const el = render(
-    <LocationProvider
-      history={
-        ({
-          listen: jest.fn(),
-          location: { search } as Location
-        } as unknown) as History
-      }
-    >
-      <Component />
-    </LocationProvider>
-  );
-
-  await tick();
-  await waitForElement(() => el.container.querySelector('a'));
-
-  const a = el.container.querySelector('a');
-  return a ? a.getAttribute('href') : '';
-}
 
 beforeAll(() => {
   jest.spyOn(console, 'error').mockImplementation(() => null);
@@ -65,11 +39,11 @@ test('DiscoverTransactionLink should produce the correct URL', async () => {
     }
   } as Transaction;
 
-  const href = await renderLink(
+  const href = await getRenderedHref(
     () => <DiscoverTransactionLink transaction={transaction} />,
     {
       search: '?rangeFrom=now/w&rangeTo=now'
-    }
+    } as Location
   );
 
   expect(href).toEqual(
@@ -84,9 +58,9 @@ test('DiscoverSpanLink should produce the correct URL', async () => {
     }
   } as Span;
 
-  const href = await renderLink(() => <DiscoverSpanLink span={span} />, {
+  const href = await getRenderedHref(() => <DiscoverSpanLink span={span} />, {
     search: '?rangeFrom=now/w&rangeTo=now'
-  });
+  } as Location);
 
   expect(href).toEqual(
     `/app/kibana#/discover?_g=(refreshInterval:(pause:true,value:'0'),time:(from:now%2Fw,to:now))&_a=(index:apm-index-pattern-id,interval:auto,query:(language:lucene,query:'span.id:"test-span-id"'))`
@@ -102,9 +76,12 @@ test('DiscoverErrorLink should produce the correct URL', async () => {
       grouping_key: 'grouping-key'
     }
   } as APMError;
-  const href = await renderLink(() => <DiscoverErrorLink error={error} />, {
-    search: '?rangeFrom=now/w&rangeTo=now'
-  });
+  const href = await getRenderedHref(
+    () => <DiscoverErrorLink error={error} />,
+    {
+      search: '?rangeFrom=now/w&rangeTo=now'
+    } as Location
+  );
 
   expect(href).toEqual(
     `/app/kibana#/discover?_g=(refreshInterval:(pause:true,value:'0'),time:(from:now%2Fw,to:now))&_a=(index:apm-index-pattern-id,interval:auto,query:(language:lucene,query:'service.name:"service-name" AND error.grouping_key:"grouping-key"'),sort:('@timestamp':desc))`
@@ -121,11 +98,11 @@ test('DiscoverErrorLink should include optional kuery string in URL', async () =
     }
   } as APMError;
 
-  const href = await renderLink(
+  const href = await getRenderedHref(
     () => <DiscoverErrorLink error={error} kuery="some:kuery-string" />,
     {
       search: '?rangeFrom=now/w&rangeTo=now'
-    }
+    } as Location
   );
 
   expect(href).toEqual(
